major refactor. import statements mostly fixed.
