+++ src/gui/tabs/tab_01_dashboard.py
