+++ src/gui/tabs/tab_02_analyze_history.py
@@ -11,13 +11,13 @@ from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
 import matplotlib.pyplot as plt
 
 # import Statement_Classes
-from Statement_Classes import Ledger
+from statement_types import Ledger
 
 # import user defined helper modules
-from Finance_GUI import gui_helper
-from src.categories import helper
-from analyzing import analyzer_helper
-from analyzing import graphing_analyzer
+from gui import gui_helper
+from categories import categories_helper
+from analysis import analyzer_helper
+from analysis import graphing_analyzer
 from db import db_helper
 
 from tools import date_helper
@@ -224,7 +224,7 @@ class tabSpendingHistory:
         fr_g_ledger.grid(row=1, column=2)
 
         ### get pyplot figure, patches, and texts
-        figure, categories, amounts = graphing_analyzer.create_pie_chart(self.transactions, helper.load_categories(), printmode="debug")
+        figure, categories, amounts = graphing_analyzer.create_pie_chart(self.transactions, gui_helper.load_categories(), printmode="debug")
 
         # add figure to FigureCanvas
         canvas = FigureCanvasTkAgg(figure, fr_graph)
@@ -267,7 +267,7 @@ class tabSpendingHistory:
         self.check_data_load_status()
 
         # get pyplot figure
-        figure = graphing_analyzer.create_top_pie_chart(self.transactions, helper.load_categories())
+        figure = graphing_analyzer.create_top_pie_chart(self.transactions, gui_helper.load_categories())
         canvas = FigureCanvasTkAgg(figure, self.fr_graph_proc)
         canvas.get_tk_widget().grid(row=2, column=0)
 
@@ -280,7 +280,7 @@ class tabSpendingHistory:
         self.check_data_load_status()
 
         # get pyplot figure
-        figure = graphing_analyzer.create_top_pie_chart(self.transactions, helper.load_categories())
+        figure = graphing_analyzer.create_top_pie_chart(self.transactions, gui_helper.load_categories())
         canvas = FigureCanvasTkAgg(figure, self.fr_graph_proc)
         canvas.get_tk_widget().grid(row=2, column=0)
