+++ src/statement_types/Robinhood.py
@@ -1,8 +1,7 @@
-from Statement_Classes import Statement
-from Statement_Classes import Transaction
+import statement_types.Statement as Statement
+import statement_types.Transaction as Transaction
 
-from Finance_GUI import gui_helper
-
+from gui import gui_helper
 from PyPDF2 import PdfFileReader
