+++ src/statement_types/VanguardRoth.py
@@ -1,8 +1,8 @@
 
 # import needed modules
-from Statement_Classes import Statement
-from Statement_Classes import Transaction
-from Finance_GUI import gui_helper
+import statement_types.Statement as Statement
+import statement_types.Transaction as Transaction
+from gui import gui_helper
 from PyPDF2 import PdfFileReader
