+++ src/statement_types/WellsSaving.py
@@ -1,9 +1,9 @@
 import csv
 
-from Statement_Classes import Statement
-from Statement_Classes import Transaction
+import statement_types.Statement as Statement
+import statement_types.Transaction as Transaction
 
-from Finance_GUI import gui_helper
+from gui import gui_helper
 
 
 class WellsSaving(Statement.Statement):
